requirements.txt
apache-airflow
apache-airflow-providers-postgres
pandas
sqlalchemy
psycopg2-binary
scikit-learn
numpy